requirements.txt
uvicorn==0.21.1
fastapi==0.95.0
httpx==0.23.3
pyjwt==2.6.0
python-dotenv==0.21.1
python-multipart==0.0.6
pytest==7.2.2
psycopg2-binary==2.9.5
SQLAlchemy==2.0.7
alembic==1.10.2
pytest-cov==4.0.0
requests==2.28.2
SQLAlchemy-Utils==0.40.0
phonenumbers==8.13.8
pydantic[email]==1.10.7
passlib[bcrypt]==1.7.4
gunicorn==20.1.0
boto3==1.26.108
fastapi-filter[sqlalchemy]==0.5.4